feat: getComponentEntities
